adding page component
