+++ app/components/views/SecurityPage/FormSelector.js
@@ -6,7 +6,7 @@ import "style/SecurityCenterMessagePage.less";
 import SignMessageForm from "./Forms/SignMessageForm";
 import VerifyMessageForm from "./Forms/VerifyMessageForm";
 
-const Page = ({
+const FormSelector = ({
   ...props,
   form,
   onSetForm,
@@ -19,46 +19,41 @@ const Page = ({
   const { signMessageSuccess, signMessageError, verifyMessageSuccess, verifyMessageError } = props;
 
   return (
-    <div className="page-view">
-      <Header
-        headerTitleOverview={<T id="invalidRPCVersion.title" m="Invalid RPC Version" />}
-      />
-      <div className="page-content message message-sign">
-        <div className="security-page-toggle">
-          <div className="text-toggle">
-            <div className={`text-toggle-button-left ${form === 0 ? "text-toggle-button-active" : null}`}
-              onClick={()=>onSetForm(0)}
-            >
-              <T id="securitycenter.header.toggle.sign" m="Sign" />
-            </div>
-            <div className={`text-toggle-button-right ${form === 1 ? "text-toggle-button-active" : null}`}
-              onClick={()=>onSetForm(1)}
+    <Aux>
+      <div className="security-page-toggle">
+        <div className="text-toggle">
+          <div className={`text-toggle-button-left ${form === 0 ? "text-toggle-button-active" : null}`}
+            onClick={() => onSetForm(0)}
+          >
+            <T id="securitycenter.header.toggle.sign" m="Sign" />
+          </div>
+          <div className={`text-toggle-button-right ${form === 1 ? "text-toggle-button-active" : null}`}
+            onClick={() => onSetForm(1)}
 
-            >
-              <T id="securitycenter.header.toggle.verify" m="Verify" />
-            </div>
+          >
+            <T id="securitycenter.header.toggle.verify" m="Verify" />
           </div>
         </div>
-        {
-          form === 0 ? <SignMessageForm {...{
-            formatMessage,
-            onSubmitSignMessage,
-            onShowSignMessageInfo,
-            signMessageSuccess,
-            signMessageError,
-          }} /> :
-            form === 1 ? <VerifyMessageForm {...{
-              formatMessage,
-              onSubmitVerifyMessage,
-              onShowVerifyMessageInfo,
-              verifyMessageSuccess,
-              verifyMessageError,
-            }} />
-              : null
-        }
       </div>
-    </div>
+      {
+        form === 0 ? <SignMessageForm {...{
+          formatMessage,
+          onSubmitSignMessage,
+          onShowSignMessageInfo,
+          signMessageSuccess,
+          signMessageError,
+        }} /> :
+          form === 1 ? <VerifyMessageForm {...{
+            formatMessage,
+            onSubmitVerifyMessage,
+            onShowVerifyMessageInfo,
+            verifyMessageSuccess,
+            verifyMessageError,
+          }} />
+            : null
+      }
+    </Aux>
   );
 };
 
-export default Page;
+export default FormSelector;
