+++ app/connectors/index.js
@@ -17,14 +17,12 @@ export { default as send } from "./send";
 export { default as service } from "./service";
 export { default as settings } from "./settings";
 export { default as sideBar } from "./sideBar";
-export { default as signMessagePage } from "./signMessagePage";
 export { default as stakeInfo } from "./stakeInfo";
 export { default as stakePools } from "./stakePools";
 export { default as ticketAutoBuyer } from "./ticketAutoBuyer";
 export { default as ticketsPage } from "./ticketsPage";
 export { default as transactionDetails } from "./transactionDetails";
 export { default as transactionPage } from "./transactionPage";
-export { default as verifyMessagePage } from "./verifyMessagePage";
 export { default as votingPrefs } from "./votingPrefs";
 export { default as walletError } from "./walletError";
 export { default as walletStartup } from "./walletStartup";
